fix dark mode error
