+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "next": "11.1.2",
+    "next-dark-mode": "^3.0.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "styled-components": "^5.3.1"
