+++ pages/_app.js
@@ -1,7 +1,8 @@
+import withDarkMode, { MODE } from 'next-dark-mode'
 import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />
 }
 
-export default MyApp
+export default withDarkMode(MyApp, { defaultMode: MODE.LIGHT })
